requirement.txt
pandas
tweepy
azure-eventhub
azure-eventhub-checkpointstoreblob-aio
azure-storage-blob
avro-python3
